fix: implement error.tsx + loading.tsx in home page
